More fixups and code reorg.
