+++ wsproxy/service/main.py
@@ -20,6 +20,7 @@ from wsproxy.protocol.json import once, setup_subscription
 from wsproxy.routes import registry as route_registry
 from wsproxy.authentication.manager import BasicPasswordAuthFactory
 from wsproxy.service.config import get_default_config_file_contents
+from wsproxy.service.parser import run_with_options, parse_auth_manager
 
 
 class UnixResolver(netutil.Resolver):
@@ -88,7 +89,7 @@ class AdminContext(object):
         self.config_options = config_options if config_options else dict()
 
     def get_connection(self):
-        auth_manager = _parse_auth_manager(self.config_options)
+        auth_manager = parse_auth_manager(self.config_options)
         if not auth_manager:
             raise Exception('Error: no authentication credentials parsed!')
         url = self.url
@@ -255,163 +256,6 @@ def socks5_command(admin_context, socks_port, to_cxn_id):
         logger.exception("Error running command!")
     finally:
         loop.close()
-
-
-def _create_admin_socket(context, path=None):
-    path = path or '/tmp/wsproxy.sock.{}'.format(os.getpid())
-
-    routes = [
-        (r'/ws', core.WsServerHandler, dict(context=context)),
-    ]
-
-    logger.info("Running admin socket on UNIX path: %s", path)
-    unix_app = web.Application(routes)
-    unix_server = httpserver.HTTPServer(unix_app)
-    unix_socket = netutil.bind_unix_socket(path)
-    unix_server.add_sockets([unix_socket])
-    unix_server.start()
-    return unix_server
-
-
-def _parse_auth_manager(options):
-    """Parse out the auth manager for a particular set of options."""
-    manager = None
-    auth_options = options.get('auth', dict())
-    if not auth_options:
-        return None
-
-    auth_type = auth_options.get('type', 'basic').lower()
-
-    # By default, do not permit localhost or any private subnets. This options
-    # will override the other filters explicitly.
-    permit_localhost = auth_options.get('permit_localhost', False)
-    permit_private_subnets = auth_options.get('permit_private_subnets', False)
-    # TODO -- Parse out explicit IPs and ports
-    kwargs = dict(
-        permit_localhost=permit_localhost,
-        permit_private_subnets=permit_private_subnets)
-
-    if auth_type == 'basic':
-        subject = auth_options.get('username')
-        password = auth_options.get('password')
-        manager = auth.BasicPasswordAuthManager(subject, password, **kwargs)
-
-    return manager
-
-
-def _parse_server_options(context, server_options):
-    servers = []
-    # TODO -- permit configuring which routes are allowed.
-    port = server_options.get('port', 8080)
-    logger.info("Running server on port: %d", port)
-
-    ssl_options = server_options.get('ssl', {})
-    if ssl_options.get('enabled', False):
-        import ssl
-
-        cert_path = ssl_options['cert_path']
-        key_path = ssl_options['key_path']
-
-        ssl_ctx = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
-        ssl_ctx.load_cert_chain(cert_path, key_path)
-        logger.info("Enabling SSL with cert: %s", cert_path)
-    else:
-        ssl_ctx = None
-        logger.warning("SSL DISABLED")
-
-    # Parse the permitted clients for this server.
-    for client in server_options.get('clients', []):
-        manager = _parse_auth_manager(client)
-        if not manager:
-            continue
-        context.auth_context.add_auth_manager(manager)
-
-    routes = [
-        (r'/ws', core.WsServerHandler, dict(context=context)),
-    ]
-    port_app = web.Application(routes)
-    port_server = httpserver.HTTPServer(port_app, ssl_options=ssl_ctx)
-    sockets = netutil.bind_sockets(port)
-    port_server.add_sockets(sockets)
-    port_server.start()
-    servers.append(port_server)
-
-    return servers
-
-
-def _parse_client_options(context, client_options):
-    url = client_options['url']
-    logger.info('Client will connect to URL: %s', url)
-    custom_ssl = client_options.get('ssl', {})
-    if custom_ssl.get('enabled', False):
-        cert_path = custom_ssl.get('cert_path')
-        # Assume true by default.
-        verify_host = custom_ssl.get('verify_host', True)
-    else:
-        cert_path = None
-
-    if cert_path:
-        import ssl
-
-        ssl_context = ssl.SSLContext(ssl.PROTOCOL_TLS)
-        ssl_context.verify_mode = ssl.CERT_REQUIRED
-        ssl_context.check_hostname = verify_host
-        ssl_context.load_verify_locations(cert_path)
-        logger.info('Using cert chain: %s (verify host: %s)',
-                    cert_path, 'True' if verify_host else 'False')
-    else:
-        ssl_context = None
-
-    request = httpclient.HTTPRequest(url, ssl_options=ssl_context)
-    return core.WsClientConnection(context, request)
-
-
-def run_with_options(options):
-    debug = options.get('debug', 0)
-    # These are the main username/password fields for this user, not for
-    # other users that could authenticate with this instance.
-    auth_manager = _parse_auth_manager(options)
-    if not auth_manager:
-        raise Exception("No root user configured!")
-    auth_context = auth.AuthContext(auth_manager)
-    auth_context.add_auth_manager(auth_manager)
-
-    # Setup the routes.
-    route_mapping = route_registry.get_route_mapping()
-    context = core.WsContext(auth_context, route_mapping, debug=debug)
-
-    # Parse server options.
-    servers = []
-    server_options = options.get('server', {})
-    if server_options.get('enabled', False):
-        servers = _parse_server_options(context, server_options)
-
-    clients = []
-    for client_options in options.get('clients', []):
-        client = _parse_client_options(context, client_options)
-        clients.append(client)
-
-    # Exit early if there is nothing to run.
-    if not servers and not clients:
-        logger.error("No servers or clients configured to run. Exiting...")
-        return
-
-    # Otherwise, also setup the admin UNIX server socket.
-    admin_server = _create_admin_socket(context)
-    servers.append(admin_server)
-
-    # Run with all of the options now passed.
-    try:
-        loop = ioloop.IOLoop.current()
-        # Queue up all of the clients to run.
-        for client in clients:
-            loop.add_callback(client.run)
-
-        # Start the loop.
-        loop.start()
-        loop.close()
-    except Exception:
-        logger.exception('Error running event loop!')
 
 
 if __name__ == '__main__':
